Initial commit, works a bit.
Very buggy at the moment, but supports Google Bookmarks and Opera Link.
Moving folders inside Chrome works for Google bookmarks, but not for
Opera Link. Adding and deleting folders and bookmarks is supported by
both. (re)moving folders externally will copy bookmarks to the old
location. It should not remove bookmars at the moment, but it may cause
a lot of mess in your bookmarks.