Remove empty folders in Google Bookmarks
Removed/renamed folders in Google bookmarks will now be moved too, but
it requires multiple synchronizations to fully complete this. It gives
errors that keeps it from doing all it should do.
Because Google Chrome doesn't report the full error message, it is hard
to debug where exactly the error is caused.