Fixed errors when renamin Google bookmark labels
This was a bug that was not trivial to solve. The gbm link used it's own
data instead of the global data (worked when gbm was the base, but not
now Google Chrome has become the base).

FIXME: the second sync after a label has been renamed, that folder is
removed and re-created (without errors - you would hardly notice it
without looking at the log).