Version to 0.2 (Opera Link support); files moved; files added to XPCOM
Updated version to 0.2. I think the Opera Link support is now at a
degree that it deserves a version increase.

A lot of files are moved to chrome/content, to support chrome:// URLs in
Firefox. This was needed for...

The XPCOM component now loads all files needed, so it now needs a way to
interact with the Places API.