Lots of fixes in Firefox and Google Bookmarks
The Firefox version is now truly working! It has very limited
functionality and (I think) lots of bugs and the only working link is
Google Bookmarks, but it works.
There are a few bugs fixed in Google Bookmarks support and the sync
engine in general (mainly to prevent strange bugs in the future).