Fixed bugs and added more error reporting
Fixed a bug where properties (like opl_id) were not copied from the new
(extern) folder to the browser data structure.

Modified code so save_state will only be called when both the link and
the browser are ready. Before this change, the browser wasn't always
ready when the link was and caused strange bugs this way.

Added error-checking code to prevent wrong status information to be
saved in Opera Link. It warns when the data is not complete and outputs
the bugus object.