Link start code cleaned up; bugfix for opera
Cleaned up link.start: that function has been split in two separate
functions, link.start and link.enable.
link.start (re)starts the link and is different per link
link.enable enables the link (moved from init and start to that single
function) and starts it, but does nothing when it is not
enabled.

Bugfix for opera: now operalink.js can open a new tab for authentication
when needed.