Did a very large refactor and fixed a few bugs
Moved a lot of code from google-bookmarks.js to tagtree.js, thereby
letting gbm do what it should do: parsing bookmarks and uploading
changes. It is still not finished.

Also did a few other things, but those are too much to enumerate here.
For example, notifications now don't go to finished links, but to links
that want to hear notifications.