Downloading Firefox Sync bookmarks working!!!
Just as the title says. I still have a lot to do before it really works,
but it is a start. For example, uploading bookmarks must be done and
finding some way to only report changes in Firefox Sync (that will be
much more efficient, because the bookmarks are slow to download).